win: spawn: correct return type

+++ lib/plat/windows/windows-spawn.c
@@ -579,5 +579,5 @@ lws_spawn_get_fd_stdxxx(struct lws_spawn_piped *lsp, int std_idx)
 {
 	assert(std_idx >= 0 && std_idx < 3);
 
-	return lsp->pipe_fds[std_idx][!!(std_idx == 0)];
+	return (int)(intptr_t)lsp->pipe_fds[std_idx][!!(std_idx == 0)];
 }
